+++ mrb_doc/models/image.rb
@@ -171,6 +171,13 @@ class Image
     self
   end
 
+  # Inverts the colours of the image
+  # @return [nil]
+  def invert!
+    # src/mruby_integration/models/image.cpp
+    self
+  end
+
   # Generates a new image of width by height in the specified colour.
   # @param width [Integer]
   # @param height [Integer]