+++ test/fixtures/models/image.rb
@@ -126,3 +126,9 @@ def fixture_models_image_colour_tint!
     Colour.new(0, 108, 45, 255)
   ]
 end
+
+def fixture_models_image_colour_invert!
+  [
+    WHITE
+  ]
+end