+++ test/models/image.rb
@@ -204,4 +204,13 @@ class TestImage < MTest::Unit::TestCase
 
     image.unload
   end
+
+  def test_image_invert!
+    image = Image.generate(width: 1, height: 1, colour: BLACK)
+
+    image.invert!
+    assert_equal fixture_models_image_colour_invert!, image.data
+
+    image.unload
+  end
 end